+++ src/store/AddLeadStore.tsx
@@ -1,8 +1,14 @@
+import axios from 'axios';
 import { useReducer } from 'react';
 
 import { useToast } from '@chakra-ui/react';
+import APIEndpoints from '../utils/endpoints';
 
-export const useAddLeadStore = () => {
+export const useAddLeadStore = (): AddLeadStore => {
+  // Initialise toast
+  // ? Maybe this could be initialised only in the main ?
+  const toast = useToast({ position: 'top-right', variant: 'solid' });
+
   const initialState: State = {
     data: {
       firstName: '',
@@ -18,62 +24,100 @@ export const useAddLeadStore = () => {
   };
 
   /// Return the error or null
-  const isValidState = (state: State): null | string => {
+  const validateForm = (state: State): null | string => {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const error = Object.entries(state.data).find(([_, value]) => value === '');
-    return error ? error[1] : null;
+    // Return the key of the first pair [key,value] to have an empty value
+    return error ? error[0] : null;
   };
 
-  const reportError = (errorMessage: string): void => {
+  // Notify the error via toasts
+  const notifyError = (errorMessage: string): void => {
     toast({
       title: "Impossible d'enregister le contact",
       description: errorMessage,
       status: 'error',
+      isClosable: true,
     });
   };
-  const toast = useToast({ position: 'top-right', variant: 'solid' });
 
-  /// Save contact
+  // Save contact
   const saveNewLead = async (): Promise<boolean> => {
     dispatch({ type: ActionType.SetIsLoading, payload: true });
-    const maybeError = isValidState(state);
+
+    const maybeError = validateForm(state);
 
     if (maybeError) {
-      reportError(maybeError);
-      // dispatch({ type: ActionType.SetIsLoading, payload: false });
-      return false;
+      notifyError(maybeError);
+      dispatch({ type: ActionType.SetIsLoading, payload: false });
+      // Wrap the false into a promise before returning
+      return new Promise((res) => res(false));
     }
 
+    // Filter the empty values out
     const body = Object.fromEntries(
       Object.entries(state.data).filter(([_, value]) => value !== '')
     );
 
-    //! Note should be optional...
-    // const res = await axios.post('http://127.0.0.1:8080/lead', {
-    //   ...body,
-    //   note: 'Hello',
-    // });
-    //  body);
-    // dispatch({ type: ActionType.SetIsLoading, payload: false });
-    return true;
+    // //! Note should be optional...
+    return axios
+      .post(APIEndpoints.lead.add, {
+        ...body,
+        note: 'Hello',
+      })
+      .then(() => {
+        resetForm();
+        toast({
+          status: 'success',
+          title: "C'est tout bon!",
+          description: `${state.data.firstName} a été sauvegardé`,
+          isClosable: true,
+        });
+        dispatch({ type: ActionType.SetIsLoading, payload: false });
+        return true;
+      })
+      .catch((e) => {
+        notifyError(e.toString());
+        dispatch({ type: ActionType.SetIsLoading, payload: false });
+        return false;
+      });
   };
+
+  function resetForm() {
+    dispatch({ type: ActionType.ResetForm, payload: null });
+  }
 
   const formReducer = (state: State, action: Action): State => {
     const { type, payload } = action;
     switch (type) {
       case ActionType.ChangeEmail:
-        return { ...state, data: { ...state.data, email: payload } };
+        return { ...state, data: { ...state.data, email: payload as string } };
       case ActionType.ChangeFirstName:
-        return { ...state, data: { ...state.data, firstName: payload } };
+        return {
+          ...state,
+          data: { ...state.data, firstName: payload as string },
+        };
       case ActionType.ChangePhone:
-        return { ...state, data: { ...state.data, phoneNumber: payload } };
+        return {
+          ...state,
+          data: { ...state.data, phoneNumber: payload as string },
+        };
       case ActionType.ChangeTitle:
-        return { ...state, data: { ...state.data, title: payload } };
+        return { ...state, data: { ...state.data, title: payload as string } };
       case ActionType.ChangeLastName:
-        return { ...state, data: { ...state.data, lastName: payload } };
+        return {
+          ...state,
+          data: { ...state.data, lastName: payload as string },
+        };
       case ActionType.ChangeCompany:
-        return { ...state, data: { ...state.data, companyId: payload } };
+        return {
+          ...state,
+          data: { ...state.data, companyId: payload as string },
+        };
       case ActionType.SetIsLoading:
-        return { ...state, isLoading: payload };
+        return { ...state, isLoading: payload as boolean };
+      case ActionType.ResetForm:
+        return initialState;
       default:
         return state;
     }
@@ -103,10 +147,19 @@ export type State = {
   isLoading: boolean;
 };
 
-export type Action = {
-  type: ActionType;
-  payload: string;
-};
+export type Action =
+  | {
+      type:
+        | ActionType.ChangeCompany
+        | ActionType.ChangeEmail
+        | ActionType.ChangeFirstName
+        | ActionType.ChangeLastName
+        | ActionType.ChangeTitle
+        | ActionType.ChangePhone;
+      payload: string;
+    }
+  | { type: ActionType.SetIsLoading; payload: boolean }
+  | { type: ActionType.ResetForm; payload: null };
 
 export enum ActionType {
   ChangeFirstName = 'CHANGE_FIRST_NAME',
@@ -116,4 +169,11 @@ export enum ActionType {
   ChangePhone = 'CHANGE_PHONE',
   ChangeCompany = 'CHANGE_COMPANY',
   SetIsLoading = 'SET_IS_LOADING',
+  ResetForm = 'RESET_FORM',
 }
+
+type AddLeadStore = {
+  state: State;
+  dispatch: React.Dispatch<Action>;
+  saveNewLead: () => Promise<boolean>;
+};