+++ src/utils/endpoints.ts
@@ -1,0 +1,10 @@
+const APIEndpoints = {
+  lead: {
+    add: 'http://127.0.0.1:8080/lead',
+  },
+  companies: {
+    get: 'http://127.0.0.1:8080/company',
+  },
+};
+
+export default APIEndpoints;