+++ src/utils/queryKeys.ts
@@ -1,0 +1,5 @@
+enum QueryKeys {
+  Companies = 'COMPANIES',
+}
+
+export default QueryKeys;